add proc-macro for generating kotlin types
