+++ Cargo.toml
@@ -1,40 +1,7 @@
-[package]
-name = "mutant-kraken"
-version = "0.1.2"
-edition = "2021"
-license = "MIT"
-authors = ["Josue Molina Morales <molinajosue92@hotmail.com>"]
-description = "A mutation testing tool for Kotlin"
-repository = "https://github.com/JosueMolinaMorales/mutant-kraken"
-keywords = ["mutation", "testing", "kotlin"]
+[workspace]
+members = [
+    "mutant-kraken",
+    "mutant-kraken-kt-types-macro"
+]
 
-[[bin]]
-name = "mutant-kraken"
-path = "src/main.rs"
-
-[lib]
-path = "src/lib.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.1.6", features = ["derive"] }
-cli-table = "0.4.7"
-csv = "1.2.1"
-indicatif = "0.17.3"
-rand = "0.8.5"
-rayon = "1.7.0"
-serde_json = "1.0.93"
-serde = { version = "1", features = ["derive"] }
-tracing = "0.1.37"
-tracing-subscriber = "0.3.16"
-tree-sitter = "0.20.9"
-tree-sitter-kotlin = "0.2.11"
-uuid = { version = "1.3.0", features = ["v4", "serde"] }
-wait-timeout = "0.2.0"
-regex = "1.8.1"
-tracing-appender = "0.2.2"
-horrorshow = "0.8.4"
-num-traits = "0.2.17"
-assert_cmd = "2.0.12"
-predicates = "3.0.4"
+resolver = "1"
