+++ mutant-kraken/src/error.rs
