+++ mutant-kraken/src/html_gen.rs
