+++ mutant-kraken/src/main.rs
@@ -1,5 +1,4 @@
 use mutant_kraken::cli::run_cli;
-
 fn main() {
     run_cli();
 }
