+++ mutant-kraken/tests/kotlin-test-projects/demo/build.gradle.kts
