+++ mutant-kraken/tests/kotlin-test-projects/demo/gradle.properties
