+++ mutant-kraken/tests/kotlin-test-projects/demo/gradle/wrapper/gradle-wrapper.properties
