+++ mutant-kraken/tests/kotlin-test-projects/demo/settings.gradle.kts
