+++ mutant-kraken/tests/kotlin-test-projects/demo/src/main/kotlin/NumberOperations.kt
