+++ mutant-kraken/tests/kotlin-test-projects/demo/src/test/kotlin/NumberOperationTest.kt
