+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/build.gradle.kts
