+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/gradle.properties
