+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/gradle/wrapper/gradle-wrapper.properties
