+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/settings.gradle.kts
