+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/src/main/kotlin/Calculator.kt
