+++ mutant-kraken/tests/kotlin-test-projects/kotlin-project/src/test/kotlin/NumberOperationTest.kt
