+++ mutant-kraken/tests/kotlin-test-projects/mutations/Killed.kt
