+++ mutant-kraken/tests/kotlin-test-projects/no-gradle-project/Calculator.kt
