+++ mutant-kraken/tests/kotlin-test-projects/no-gradle-project/Main.kt
